CLN: Fix small type mismatches on variables

+++ randomstate/distributions.h
@@ -2,10 +2,12 @@
 #include <stdlib.h>
 
 #ifdef _WIN32
+#include "src/common/stdint.h"
 typedef int bool;
 #define false 0
 #define true 1
 #else
+#include <stdint.h>
 #endif
 
 #ifndef min
@@ -121,20 +123,20 @@ extern unsigned long random_uint(aug_state* state);
 
 extern unsigned long random_interval(aug_state* state, unsigned long max);
 
-extern void random_bounded_uint64_fill(aug_state *state, uint64_t off, uint64_t rng, int cnt, uint64_t *out);
+extern void random_bounded_uint64_fill(aug_state *state, uint64_t off, uint64_t rng, intptr_t cnt, uint64_t *out);
 
-extern void random_bounded_uint32_fill(aug_state *state, uint32_t off, uint32_t rng, int cnt,uint32_t *out);
+extern void random_bounded_uint32_fill(aug_state *state, uint32_t off, uint32_t rng, intptr_t cnt, uint32_t *out);
 
-extern void random_bounded_uint16_fill(aug_state *state, uint16_t off, uint16_t rng, int cnt, uint16_t *out);
+extern void random_bounded_uint16_fill(aug_state *state, uint16_t off, uint16_t rng, intptr_t cnt, uint16_t *out);
 
-extern void random_bounded_uint8_fill(aug_state *state, uint8_t off, uint8_t rng, int cnt, uint8_t *out);
+extern void random_bounded_uint8_fill(aug_state *state, uint8_t off, uint8_t rng, intptr_t cnt, uint8_t *out);
 
-extern void random_bool_fill(aug_state *state, int8_t off, int8_t rng, int cnt, int8_t *out);
+extern void random_bool_fill(aug_state *state, int8_t off, int8_t rng, intptr_t cnt, int8_t *out);
 
-extern void random_uniform_fill(aug_state* state, int count, double *out);
+extern void random_uniform_fill(aug_state* state, intptr_t count, double *out);
 
-extern void random_standard_exponential_fill(aug_state* state, int count, double *out);
+extern void random_standard_exponential_fill(aug_state* state, intptr_t count, double *out);
 
-extern void random_gauss_fill(aug_state* state, int count, double *out);
+extern void random_gauss_fill(aug_state* state, intptr_t count, double *out);
 
-extern void random_gauss_zig_julia_fill(aug_state *state, int count, double *out);
+extern void random_gauss_zig_julia_fill(aug_state *state, intptr_t count, double *out);
